refactor(git): split 2495-line git.rs into focused submodules
